requirements.txt
amqp==5.3.1
asgiref==3.11.1
azure-ai-formrecognizer==3.3.3
azure-common==1.1.28
azure-core==1.38.0
billiard==4.2.4
boto3==1.35.0
botocore==1.35.99
celery==5.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.1
django-allauth==64.2.1
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
idna==3.11
ImageHash==4.3.1
isodate==0.7.2
jmespath==1.1.0
kombu==5.6.2
msrest==0.7.1
numpy==2.4.2
oauthlib==3.3.1
opencv-python-headless==4.10.0.84
packaging==26.0
pillow==12.1.0
prompt_toolkit==3.0.52
pypng==0.20220715.0
python-barcode==0.15.1
python-dateutil==2.9.0.post0
python-decouple==3.8
PyWavelets==1.9.0
qrcode==7.4.2
redis==5.0.8
requests==2.32.5
requests-oauthlib==2.0.0
s3transfer==0.10.4
scipy==1.17.0
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
whitenoise==6.7.0
PyMuPDF==1.24.0
openai>=1.0.0